+++ class/actes/ActesSignature.class.php
@@ -1,18 +1,48 @@
 <?php
 
+
 class ActesSignature {
 	
 	private $actesIncludedFileSQL;
 	private $actesTransactionSQL;
 	private $actesEnveloppeSQL;
+	private $actes_files_upload_root;
 
-	public function __construct($sqlQuery){
+	public function __construct(SQLQuery $sqlQuery,$actes_files_upload_root){
 		$this->actesIncludedFileSQL = new ActesIncludedFileSQL($sqlQuery);
 		$this->actesTransactionSQL = new ActesTransactionsSQL($sqlQuery);
 		$this->actesEnveloppeSQL = new ActesEnvelopeSQL($sqlQuery);
+		$this->actes_files_upload_root = $actes_files_upload_root;
 	}
-	
+
+	/**
+	 * @param $actes_included_file_id
+	 * @param $signature
+	 * @return array|bool|mixed
+	 * @throws Exception
+	 */
 	public function setSignature($actes_included_file_id,$signature){
+
+		$tmpFolder = new TmpFolder();
+		$tmp_dir = $tmpFolder->create();
+		try {
+			$result = $this->setSignatureThrowException($actes_included_file_id, $signature,$tmp_dir);
+		} catch (Exception $e){
+			$tmpFolder->delete($tmp_dir);
+			throw $e;
+		}
+		$tmpFolder->delete($tmp_dir);
+		return $result;
+	}
+
+	/**
+	 * @param $actes_included_file_id
+	 * @param $signature
+	 * @param $tmp_dir
+	 * @return array|bool|mixed
+	 * @throws Exception
+	 */
+	public function setSignatureThrowException($actes_included_file_id,$signature,$tmp_dir){
 		$transaction_id = $this->actesIncludedFileSQL->getTransactionId($actes_included_file_id);
 		if (! $transaction_id){
 			throw new Exception("Impossible de trouver une transaction ratachée au fichier à signé");
@@ -26,34 +56,32 @@ class ActesSignature {
 			throw new Exception("La transaction $transaction_id n'est pas dans l'état « En attente d'être signée. »");
 		}
 
-		$tmpFolder = new TmpFolder();
-		$tmpDir = $tmpFolder->create();
 		
 		$envelope_id = $transactionInfo['envelope_id'];
 		
 		
 		$actes_envelope_info = $this->actesEnveloppeSQL->getInfo($envelope_id);
-		$archivePath = ACTES_FILES_UPLOAD_ROOT.'/'.$actes_envelope_info['file_path'];
+		$archivePath = $this->actes_files_upload_root.'/'.$actes_envelope_info['file_path'];
 		
-		$tgzExtractor = new TGZExtractor($tmpDir);
+		$tgzExtractor = new TGZExtractor($tmp_dir);
 		$tgzExtractor->extract($archivePath, false);
 		
 		$xml_file = $this->actesIncludedFileSQL->getXMLFilename($transaction_id);
 		
-		$xml = simplexml_load_file($tmpDir."/".$xml_file);
+		$xml = simplexml_load_file($tmp_dir."/".$xml_file);
 		$namespaces = $xml->getDocNamespaces();
 		
 		$children = $xml->children($namespaces['actes']);
-		$children->Document->addChild("Signature",$signature,$namespaces['actes']);
+		$children->{'Document'}->addChild("Signature",$signature,$namespaces['actes']);
 		
-		$xml->asXML($tmpDir."/".$xml_file);
-		
-		chdir($tmpDir);
+		$xml->asXML($tmp_dir."/".$xml_file);
+
+		$old_cvd = getcwd();
+		chdir($tmp_dir);
 		$cmd = "tar cf - * | gzip -9 > $archivePath ";
-		
-		$status = system($cmd, $ret);
-		
-		$tmpFolder->delete($tmpDir);
+
+		system($cmd, $ret);
+		chdir($old_cvd);
 		$this->actesIncludedFileSQL->setSignature($transaction_id,$actes_included_file_id,$signature);
 		$this->actesTransactionSQL->updateStatus($transaction_id,1, "L'acte a été signé électroniquement");
 		return $transaction_id;