+++ model/AuthoritySiretSQL.class.php
@@ -3,10 +3,16 @@
 class AuthoritySiretSQL extends SQL {
 	
 	public function siretList($authority_id){
-		$sql = "SELECT * FROM authority_siret WHERE authority_id=? ORDER BY siret";
+		$sql = "SELECT * FROM authority_siret WHERE authority_id=? and is_blocked=FALSE ORDER BY siret";
 		return $this->query($sql,$authority_id);
 	}
-	
+
+
+	public function siretListBlocked($authority_id){
+		$sql = "SELECT * FROM authority_siret WHERE authority_id=? and is_blocked=TRUE ORDER BY siret";
+		return $this->query($sql,$authority_id);
+	}
+
 	public function add($authority_id,$siret){
 		$sql = "SELECT id FROM authority_siret WHERE authority_id=? AND siret=?";
 		$id = $this->queryOne($sql,$authority_id,$siret);
@@ -23,7 +29,7 @@ class AuthoritySiretSQL extends SQL {
 	}
 	
 	public function authorityList($siret){
-		$sql = "SELECT * FROM authority_siret WHERE siret=?";
+		$sql = "SELECT * FROM authority_siret WHERE siret=? AND is_blocked=FALSE";
 		return $this->query($sql,$siret);
 	}
 	
@@ -31,5 +37,16 @@ class AuthoritySiretSQL extends SQL {
 		$sql = "SELECT * FROM authority_siret WHERE id=?";
 		return $this->queryOne($sql,$authority_siret_id);
 	}
-	
+
+	public function blocked($id){
+		$sql = "UPDATE authority_siret SET is_blocked=TRUE WHERE id=?";
+		$this->query($sql,$id);
+	}
+
+	public function unblocked($id){
+		$sql = "UPDATE authority_siret SET is_blocked=FALSE WHERE id=?";
+		$this->query($sql,$id);
+	}
+
+
 }