+++ public.ssl/modules/actes/actes_classification_request.php
@@ -3,7 +3,6 @@ require_once("../../../config/config.php");
 require_once(SITEROOT . '/class/include.class.php');
 require_once(SITEROOT . '/public.ssl/modules/actes/class/ActesEnvelope.class.php');
 require_once(SITEROOT . '/public.ssl/modules/actes/class/ActesClassification.class.php');
-require_once(dirname(__FILE__)."/class/ActesClassificationCreation.class.php");
 
 // Instanciation du module courant
 $module = new Module();