+++ test/PHPUnit/bootstrap.php
@@ -5,6 +5,10 @@ define("TRACE_FILE_PATH","/tmp/s2low-phpunit.log");
 define('HELIOS_FILES_UPLOAD_ROOT', "vfs://test/helios/");
 define("ANTIVIRUS_COMMAND","ls");
 
+define("TIMESTAMPING_CERT",__DIR__."/fixtures/timestamp_certificates/tedetis_timestamp_cert.pem");
+define("TIMESTAMPING_PRIV_KEY",__DIR__."/fixtures/timestamp_certificates/tedetis_timestamp_priv_key.pem");
+define("TIMESTAMPING_PRIV_KEY_PASS",__DIR__."/fixtures/timestamp_certificates/tedetis_timestamp_priv_key.pass");
+
 set_include_path(__DIR__."/../../ext/" . PATH_SEPARATOR .   get_include_path());
 
 require_once __DIR__."/../../docker-resources/define-from-environnement.php";